pyproject.toml
[project]
name = "eta-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "certifi>=2025.8.3",
    "dagshub>=0.6.3",
    "fastapi>=0.116.1",
    "flask>=3.1.2",
    "geopy>=2.4.1",
    "ipykernel>=6.30.1",
    "mlflow>=3.3.1",
    "numpy>=2.3.2",
    "openpyxl>=3.1.5",
    "pandas>=2.3.2",
    "pyaml>=25.7.0",
    "pymongo[srv]>=4.14.1",
    "python-dotenv>=1.1.1",
    "scikit-learn>=1.7.1",
    "scipy>=1.16.1",
    "uvicorn>=0.35.0",
    "xgboost>=3.0.4",
]

[tool.uv.workspace]
members = [
    "venv",
]
